move buildPavingManifest to road logistics
